Inventory-Risk-Explainability-main/Development-Files/APIs/UserRegistration/requirements.txt
python-dotenv
fastapi
uvicorn
httpx==0.27.2
sse_starlette
urllib3
hdbcli
cfenv
requests
pyjwt 
bcrypt
pydantic[email]
